fix "ffmpeg: protocol not found"
ffmpeg can fetch data from the Internet, and so it will look for URLs
that begin with http://, https://, ftp:// and so on. This causes
ffmpeg to fail whenever it encounters an output file name with a colon
in it and so to counteract this we prepend the output file name with
"file:" to signify that we want to store it on out local filesystem.